requirements.txt
pandas>=2.0
numpy>=1.24
plotly>=5.18
streamlit>=1.30
sqlalchemy>=2.0
psycopg2-binary>=2.9
dbt-core>=1.7
dbt-postgres>=1.7
dbt-snowflake>=1.7
matplotlib>=3.8
python-dotenv>=1.0
